suite fibinacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -30,7 +30,7 @@ void display_number(int n)
  */
 int main(void)
 {
-	double a, n, prev;
+	signed long long int a, n, prev;
 	int count;
 
 	n = 1;
